% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevcmm-package.R
\docType{package}
\name{cevcmm-package}
\alias{cevcmm}
\alias{cevcmm-package}
\title{cevcmm: Communication-Efficient Varying Coefficient Mixed-Effects Models}
\description{
Scalable inference for Varying Coefficient Mixed-Effects Models
(VCMMs) with large, correlated random effects. The package implements:
}
\details{
\itemize{
\item Sufficient-statistics (SS) iterative estimator (Algorithm 1).
\item One-step communication-efficient surrogate likelihood (CSL) estimator.
\item SVD-stabilized variants for ill-conditioned random-effect Gram matrices.
\item Kronecker and separable covariance structures for origin-destination
and group-shared random effects.
}

The package is in early development. See the project ROADMAP for the
current status.
}
\references{
Lin, L.-H. and Jalili, L. (2026). Scalable and Communication-Efficient
Varying Coefficient Mixed-Effects Models: Methodology, Theory, and
Applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lidajalili/cevcmm}
  \item Report bugs at \url{https://github.com/lidajalili/cevcmm/issues}
}

}
\author{
\strong{Maintainer}: Lida Jalili \email{lchalangarjalilideh1@gsu.edu}

Authors:
\itemize{
  \item Lida Jalili \email{lchalangarjalilideh1@gsu.edu}
  \item Li-Hsiang Lin \email{lhlin@gsu.edu}
}

}
\keyword{internal}
