% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nobs.vcmm_fit}
\alias{nobs.vcmm_fit}
\title{Number of observations from a vcmm fit}
\usage{
\method{nobs}{vcmm_fit}(object, ...)
}
\arguments{
\item{object}{A \code{vcmm_fit} object.}

\item{...}{Unused.}
}
\value{
Integer.
}
\description{
Standard \code{stats::nobs} generic. Returns \eqn{N}, the total number
of observations used to compute the fit (or the sum of node sample
sizes for fits produced via \code{\link{fit_from_summaries}}).
}
